Cleanup trigger message types
This properly separates the messages accepted by the internal
TriggerRunnerImpl from the ones accepted by TriggerRunner. This also
shows a bug/redundancy where we had a bunch of code in
TriggerRunnerImpl to handle `Stop` but as evidenced by the types now,
we never actually send this message. We send it to TriggerRunner which
then tears down the child with it.

It also shuffles around the server message type to make it clear where
it belongs to.

And of course, I managed once again to include debugging output from a
previous PR so this is now removed as well :facepalm:

changelog_begin
changelog_end
