+++ triggers/service/src/main/scala/com/digitalasset/daml/lf/engine/trigger/TriggerRunnerImpl.scala
@@ -5,7 +5,7 @@ package com.daml.lf.engine.trigger
 
 import java.util.UUID
 
-import akka.actor.typed.scaladsl.Behaviors
+import akka.actor.typed.scaladsl.{Behaviors}
 import akka.actor.typed.{ActorRef, Behavior, PostStop, PreRestart}
 import akka.stream.{KillSwitch, KillSwitches, Materializer}
 import com.daml.grpc.adapter.ExecutionSequencerFactory
@@ -24,11 +24,12 @@ import scalaz.syntax.tag._
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success}
+import TriggerRunner.{QueryingACS, Running, TriggerStatus}
 
 object TriggerRunnerImpl {
 
   final case class Config(
-      server: ActorRef[Message],
+      server: ActorRef[Server.Message],
       triggerInstance: UUID,
       party: Party,
       token: Option[String],
@@ -41,12 +42,12 @@ object TriggerRunnerImpl {
       trigger.loggingExtension + ("triggerId" -> triggerInstance.toString)
   }
 
-  import TriggerRunner._
-  // TODO Properly seal this instead of extending the one from TriggerRunner
+  sealed trait Message
   final private case class Failed(error: Throwable) extends Message
   final private case class QueryACSFailed(cause: Throwable) extends Message
   final private case class QueriedACS(runner: Runner, acs: Seq[CreatedEvent], offset: LedgerOffset)
       extends Message
+  final case class Status(replyTo: ActorRef[TriggerStatus]) extends Message
 
   private[this] val logger = ContextualizedLogger get getClass
 
@@ -59,7 +60,7 @@ object TriggerRunnerImpl {
       implicit val ec: ExecutionContext = ctx.executionContext
       val triggerInstance = config.triggerInstance
       // Report to the server that this trigger is starting.
-      config.server ! TriggerStarting(triggerInstance)
+      config.server ! Server.TriggerStarting(triggerInstance)
       logger.info(s"Trigger $name is starting")
       val appId = ApplicationId(name)
       val clientConfig = LedgerClientConfiguration(
@@ -74,83 +75,59 @@ object TriggerRunnerImpl {
 
       // Waiting for the ACS query to finish so we can build the
       // initial state.
-      def queryingACS(wasStopped: Boolean): Behaviors.Receive[Message] =
+      def queryingACS(): Behaviors.Receive[Message] =
         Behaviors.receiveMessagePartial[Message] {
           case Status(replyTo) =>
             replyTo ! QueryingACS
             Behaviors.same
           case QueryACSFailed(cause) =>
-            if (wasStopped) {
-              // The stop endpoint can't send a message to a runner
-              // that isn't in the running triggers table so this is
-              // an odd case.
-              config.server ! TriggerInitializationFailure(triggerInstance, cause.toString)
-              // However we got here though, one thing is clear. We
-              // don't want to restart the actor.
-              throw new InitializationHalted("User stopped") // Don't retry!
-            } else {
-              // Report the failure to the server.
-              config.server ! TriggerInitializationFailure(triggerInstance, cause.toString)
-              // Tell our monitor there's been a failure. The
-              // monitor's supervision strategy will respond to this
-              // (including logging the exception).
-              throw new InitializationException("Couldn't start: " + cause.toString)
-            }
+            // Report the failure to the server.
+            config.server ! Server.TriggerInitializationFailure(triggerInstance, cause.toString)
+            // Tell our monitor there's been a failure. The
+            // monitor's supervision strategy will respond to this
+            // (including logging the exception).
+            throw new InitializationException("Couldn't start: " + cause.toString)
           case QueriedACS(runner, acs, offset) =>
-            if (wasStopped) {
-              // The stop endpoint can't send a message to a runner
-              // that isn't in the running triggers table so this is
-              // an odd case.
-              config.server ! TriggerInitializationFailure(triggerInstance, "User stopped")
-              // However we got here though, one thing is clear. We
-              // don't want to restart the actor.
-              throw new InitializationHalted("User stopped") // Don't retry!
-            } else {
-              // It's possible for 'runWithACS' to throw (fail to
-              // construct a flow).
-              try {
-                // The trigger is a future that we only expect to
-                // complete if something goes wrong.
-                val (killSwitch, trigger) = runner.runWithACS(
-                  acs,
-                  offset,
-                  msgFlow = KillSwitches.single[TriggerMsg],
-                  name,
-                )
+            // It's possible for 'runWithACS' to throw (fail to
+            // construct a flow).
+            try {
+              // The trigger is a future that we only expect to
+              // complete if something goes wrong.
+              val (killSwitch, trigger) = runner.runWithACS(
+                acs,
+                offset,
+                msgFlow = KillSwitches.single[TriggerMsg],
+                name,
+              )
 
-                // If we are stopped we will end up causing the future
-                // to complete which will trigger a message that is
-                // sent to a now terminated actor. In
-                // https://doc.akka.io/docs/akka/current/general/message-delivery-reliability.html#dead-letters
-                // it is explained that this is a somewhat ordinary
-                // circumstance and not to be worried about.
-                ctx.pipeToSelf(trigger) {
-                  case Success(_) =>
-                    Failed(new RuntimeException("Trigger exited unexpectedly"))
-                  case Failure(cause) =>
-                    Failed(cause)
-                }
-                // Report to the server that this trigger is entering
-                // the running state.
-                config.server ! TriggerStarted(triggerInstance)
-                logger.info(s"Trigger $name is starting")
-                running(killSwitch)
-              } catch {
-                case cause: Throwable =>
-                  // Report the failure to the server.
-                  config.server ! TriggerInitializationFailure(triggerInstance, cause.toString)
-                  logger.info(s"Trigger $name failed during initialization: $cause")
-                  // Tell our monitor there's been a failure. The
-                  // monitor's supervisor strategy will respond to
-                  // this by writing the exception to the log and
-                  // attempting to restart this actor.
-                  throw new InitializationException("Couldn't start: " + cause.toString)
+              // If we are stopped we will end up causing the future
+              // to complete which will trigger a message that is
+              // sent to a now terminated actor. In
+              // https://doc.akka.io/docs/akka/current/general/message-delivery-reliability.html#dead-letters
+              // it is explained that this is a somewhat ordinary
+              // circumstance and not to be worried about.
+              ctx.pipeToSelf(trigger) {
+                case Success(_) =>
+                  Failed(new RuntimeException("Trigger exited unexpectedly"))
+                case Failure(cause) =>
+                  Failed(cause)
               }
+              // Report to the server that this trigger is entering
+              // the running state.
+              config.server ! Server.TriggerStarted(triggerInstance)
+              logger.info(s"Trigger $name is starting")
+              running(killSwitch)
+            } catch {
+              case cause: Throwable =>
+                // Report the failure to the server.
+                config.server ! Server.TriggerInitializationFailure(triggerInstance, cause.toString)
+                logger.info(s"Trigger $name failed during initialization: $cause")
+                // Tell our monitor there's been a failure. The
+                // monitor's supervisor strategy will respond to
+                // this by writing the exception to the log and
+                // attempting to restart this actor.
+                throw new InitializationException("Couldn't start: " + cause.toString)
             }
-          case Stop =>
-            // We got a stop message but the ACS query hasn't
-            // completed yet.
-            queryingACS(wasStopped = true)
         }
 
       // The trigger loop is running. The only thing to do now is wait
@@ -161,14 +138,9 @@ object TriggerRunnerImpl {
             case Status(replyTo) =>
               replyTo ! Running
               Behaviors.same
-            case Stop =>
-              logger.info(s"Trigger $name is stopping")
-              // Don't think about trying to send the server a message
-              // here. It won't receive it (I found out the hard way).
-              Behaviors.stopped
             case Failed(cause) =>
               // Report the failure to the server.
-              config.server ! TriggerRuntimeFailure(triggerInstance, cause.toString)
+              config.server ! Server.TriggerRuntimeFailure(triggerInstance, cause.toString)
               logger.info(s"Trigger $name failed: $cause")
               // Tell our monitor there's been a failure. The
               // monitor's supervisor strategy will respond to this by
@@ -215,6 +187,6 @@ object TriggerRunnerImpl {
         case Failure(cause) => QueryACSFailed(cause)
       }
 
-      queryingACS(wasStopped = false)
+      queryingACS()
     }
 }
