+++ triggers/service/src/test/scala/com/digitalasset/daml/lf/engine/trigger/TriggerServiceFixture.scala
@@ -453,7 +453,7 @@ trait TriggerServiceFixture
             } yield r
           } {
             case (_, system) => {
-              system ! Stop
+              system ! Server.Stop
               system.whenTerminated.map(_ => ())
             }
           }
